Implement a digest method
The digest method returns a hexadecimal string digest.  This was an
80/20 optimization in line with the Tiny philosophy to keep things
simple and return values comparable to command line tools and what
people would see on a website with a digest to verify a file.

The default algorithm is 'SHA-256' or Digest module constructors may be
given as arguments.

Those with more specific needs should just use Digest directly.