Let children() method use optional filter argument
As mentioned in GitHub #38 issue, let children() method
allow optional filter argument as `qr//` or code reference.
If invalid argument is given then children() will be croaked.
I setted code reference received two arguments, the first one
is `$self` and the second one is the name of the child to avoid
some object creation/stringification overhead.  I can't convince
`$self` is really needed or not, but it was always useful in my cases.