deprecate dirname due to inconsistencies
Because 'dirname' as used by Path::Tiny might or might not give trailing
slashes, users don't know what to expect.  Rather than changing it to
always be one thing or the other, which might break people's code that
works around it, this commit just documents the problem and notes that
the method is deprecated.

Resolves #106.