Hook for plugin system
I added a 'path_plugins' function that is doing a very simple
injection of exported functions of potential plugin modules.

The idea is to be able to add functionality to Path::Tiny without
besmirching its 'Tiny' moniker. In the added test there is a
frivolous example of a rot13 encrypter/decrypter, but I already
have something a tad more serious at https://github.com/yanick/Path-Tiny-Serialize

TBD item: I've assumed the namespace of the plugins to be
Path::Tiny::*, but you might perhaps
want to move that to Path::Tiny::Plugin::*.