change visit API for simplicty and state accumulation
The visit API now gives the callback a hashref for accumulating state
and returns it at the end; this will avoid people having to create
closures for their callbacks.

The file is aliased to $_ for simple callbacks.

The 'stop iterating' feature is changed to require a reference to
a false scalar (\0), similar to Path::Iterator::Rule. This will
help people who will forget to always return a true value.  The
default is now to always keep iterating and a special action is required
to halt.