test for symlink capability consistently
In some places the code was checking $Config{d_symlink} and in some
places it was looking for an exception when you tried to call
symlink().  In addition, check to see if the thing actually created
really is a symlink.

On MSYS2 there are no real symlinks, ln (with or without -s) just
makes copies.  I suppose because symlink() does something (though
not what it does in Unix or cygwin), $Config{d_symlink} is defined
though arguably it shouldn't (?).