Fix tilde expansion if prefixed by curdir
When a tilde appears after the current directory, like `./~xdg`, it
should not be expanded.  Previously, canonicalization stripped the
leading `./`, which then triggered tilde expansion.  This commit is more
selective about tilde expansion and also re-escapes cases like `./~xdg`
so it stays that way instead of turning into `~xdg`, which would tilde
expand if passed to Path::Tiny again.