validation: stop ignoring bogus arguments, throw exceptions
Some methods, most notably ->mkdir and ->mkpath and ->remove_tree,
expected a single hash reference (or nothing) as an argument.  If passed
a string (for example), they would treat it like {}.

This hid this broken code:

    path($root)->mkdir('name-of-child');

It would not create $root/name-of-child, but would ensure that $root
existed, then return.

This code will notice extra or non-hashref arguments and raise an
exception.  An explicit undef is permitted, although only out of a sense
of caution.